# openapi.yaml
openapi: 3.0.0
info:
  title: Indikara Backend API
  version: 1.0.0
  description: API for authentication, products, and health checks
servers:
  - url: https://backend-wei5.onrender.com
    description: Deployment server
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WebhookEvent:
      type: object
      properties:
        _id:
          type: string
        provider:
          type: string
        payload:
          type: object
        headers:
          type: object
        ip:
          type: string
        hashValid:
          type: boolean
        status:
          type: string
        receivedAt:
          type: string
          format: date-time
...
paths:
  /api/admin/webhooks:
    get:
      summary: List webhook events (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: provider
          schema:
            type: string
        - in: query
          name: hashValid
          schema:
            type: boolean
      responses:
        '200':
          description: Paginated webhook events
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookEvent'
              examples:
                sample:
                  value:
                    page: 1
                    limit: 20
                    total: 1
                    events:
                      - _id: evt_123
                        provider: payu
                        status: success
                        hashValid: true
  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
            examples:
              demo:
                value:
                  name: Alice
                  email: alice@example.com
                  password: password123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
                  token:
                    type: string
        '400':
          description: User already exists or invalid data
        '500':
          description: Server error
  /api/auth/login:
    post:
      summary: Login user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
                  token:
                    type: string
        '401':
          description: Invalid email or password
        '500':
          description: Server error
...